refactor: unify websocket endpoints with http services
