--- docs/services/tts/ws.md
@@ -1,13 +1,0 @@
-# ws.py
-
-**Path**: `services/tts/ws.py`
-
-**Description**: WebSocket interface that streams synthesized speech from text messages.
-
-## Dependencies
-- fastapi
-- soundfile
-- shared.py.speech.tts
-
-## Dependents
-- `run_ws.sh`
