--- docs/stt-ws-app.md
@@ -1,5 +1,0 @@
-# `services/stt_ws/app.py`
-
-FastAPI WebSocket server for streaming audio transcription.
-
-#hashtags: #stt #websocket #service #promethean
