--- docs/whisper-stream-ws-app.md
@@ -1,5 +1,0 @@
-# `services/whisper_stream_ws/app.py`
-
-WebSocket endpoint that consumes raw PCM chunks and sends back streaming transcriptions.
-
-#hashtags: #stt #whisper #websocket #service #promethean
