--- services/hy/whisper_stream_ws/Pipfile
@@ -1,21 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-fastapi = "*"
-numpy = "*"
-uvicorn = "*"
-hy = "*"
-
-[dev-packages]
-pytest-cov = "*"
-pytest = "*"
-flake8 = "*"
-black = "*"
-anyio = "*"
-httpx = "*"
-
-[requires]
-python_version = "3.12"
