--- services/hy/whisper_stream_ws/README.md
@@ -1,7 +1,0 @@
-# Whisper Streaming WebSocket Service
-
-Streams audio over WebSocket and returns incremental transcriptions using the new `WhisperStreamer` utility.
-
-Run with `./run.sh`.
-
-#hashtags: #stt #whisper #websocket #service #promethean
