+++ services/py/stt/README.md
@@ -1,6 +1,8 @@
 # STT Service
 
 This service handles speech-to-text using Whisper models accelerated on NPU hardware.
+It now also exposes WebSocket endpoints for single-shot transcription (`/transcribe`)
+and streaming transcription (`/stream`).
 
 ## Usage
