--- services/py/stt_ws/Pipfile
@@ -1,22 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-fastapi = "==0.116.1"
-transformers = "*"
-torch = "*"
-numpy = "*"
-uvicorn = "*"
-hy = "*"
-
-[dev-packages]
-pytest = "*"
-pytest-cov = "*"
-flake8 = "*"
-black = "*"
-httpx = "*"
-
-[requires]
-python_version = "3.12"
