--- services/py/stt_ws/README.md
@@ -1,13 +1,0 @@
-# STT WebSocket Service
-
-Exposes a WebSocket endpoint for speech-to-text transcription.
-
-## Usage
-
-Run with:
-
-```bash
-./run.sh
-```
-
-#hashtags: #stt #websocket #service #promethean
