--- services/py/stt_ws/requirements.txt
@@ -1,39 +1,0 @@
--i https://pypi.org/simple
-annotated-types==0.7.0; python_version >= '3.8'
-anyio==4.9.0; python_version >= '3.9'
-certifi==2025.8.3; python_version >= '3.7'
-charset-normalizer==3.4.2; python_version >= '3.7'
-click==8.2.1; python_version >= '3.10'
-colorama==0.4.6; platform_system == 'Windows'
-fastapi==0.116.1; python_version >= '3.8'
-filelock==3.18.0; python_version >= '3.9'
-fsspec==2025.7.0; python_version >= '3.9'
-funcparserlib==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
-h11==0.16.0; python_version >= '3.8'
-huggingface-hub==0.34.3; python_full_version >= '3.8.0'
-hy==1.1.0; python_version < '3.15' and python_version >= '3.9'
-idna==3.10; python_version >= '3.6'
-jinja2==3.1.6; python_version >= '3.7'
-markupsafe==3.0.2; python_version >= '3.9'
-mpmath==1.3.0
-networkx==3.5; python_version >= '3.11'
-numpy==2.3.2; python_version >= '3.11'
-packaging==25.0; python_version >= '3.8'
-pydantic==2.11.7; python_version >= '3.9'
-pydantic-core==2.33.2; python_version >= '3.9'
-pyyaml==6.0.2; python_version >= '3.8'
-regex==2025.7.34; python_version >= '3.9'
-requests==2.32.4; python_version >= '3.8'
-safetensors==0.5.3; python_version >= '3.7'
-setuptools==80.9.0; python_version >= '3.12'
-sniffio==1.3.1; python_version >= '3.7'
-starlette==0.47.2; python_version >= '3.9'
-sympy==1.14.0; python_version >= '3.9'
-tokenizers==0.21.4; python_version >= '3.9'
-torch==2.7.1; python_full_version >= '3.9.0'
-tqdm==4.67.1; python_version >= '3.7'
-transformers==4.54.1; python_full_version >= '3.9.0'
-typing-extensions==4.14.1; python_version >= '3.9'
-typing-inspection==0.4.1; python_version >= '3.9'
-urllib3==2.5.0; python_version >= '3.9'
-uvicorn==0.35.0; python_version >= '3.9'
