--- services/py/stt_ws/run.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/bash
-pipenv run uvicorn app:app --host 0.0.0.0 --port 5002
