+++ services/py/tts/README.md
@@ -1,6 +1,8 @@
 # TTS Service
 
 This service converts text to speech using Tacotron and WaveRNN models.
+It exposes both an HTTP endpoint (`/synth_voice_pcm`) and a WebSocket
+endpoint (`/ws/tts`) on the same server.
 
 ## Usage
