--- services/py/tts/run_ws.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/bash
-pipenv run uvicorn ws:app --host 0.0.0.0 --port 5003
