--- services/py/whisper_stream_ws/requirements.txt
@@ -1,18 +1,0 @@
--i https://pypi.org/simple
-annotated-types==0.7.0; python_version >= '3.8'
-anyio==4.9.0; python_version >= '3.9'
-click==8.2.1; python_version >= '3.10'
-colorama==0.4.6; platform_system == 'Windows'
-fastapi==0.116.1; python_version >= '3.8'
-funcparserlib==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
-h11==0.16.0; python_version >= '3.8'
-hy==1.1.0; python_version < '3.15' and python_version >= '3.9'
-idna==3.10; python_version >= '3.6'
-numpy==2.3.2; python_version >= '3.11'
-pydantic==2.11.7; python_version >= '3.9'
-pydantic-core==2.33.2; python_version >= '3.9'
-sniffio==1.3.1; python_version >= '3.7'
-starlette==0.47.2; python_version >= '3.9'
-typing-extensions==4.14.1; python_version >= '3.9'
-typing-inspection==0.4.1; python_version >= '3.9'
-uvicorn==0.35.0; python_version >= '3.9'
